+++ core/serde.fst/src/main/java/mb/pie/serde/fst/FstSerde.java
@@ -13,14 +13,20 @@ import java.io.OutputStream;
 
 public class FstSerde implements Serde {
     private final FSTConfiguration configuration;
+    private final @Nullable ClassLoader defaultClassLoader;
 
 
-    public FstSerde(FSTConfiguration configuration) {
+    public FstSerde(FSTConfiguration configuration, @Nullable ClassLoader defaultClassLoader) {
         this.configuration = configuration;
+        this.defaultClassLoader = defaultClassLoader;
+    }
+
+    public FstSerde(@Nullable ClassLoader classLoader) {
+        this(FSTConfiguration.createDefaultConfiguration(), classLoader);
     }
 
     public FstSerde() {
-        this(FSTConfiguration.createDefaultConfiguration());
+        this(null);
     }
 
 
@@ -33,7 +39,8 @@ public class FstSerde implements Serde {
         }
     }
 
-    @Override public <T> T deserialize(Class<T> type, InputStream inputStream) {
+    @Override public <T> T deserialize(Class<T> type, InputStream inputStream, @Nullable ClassLoader classLoader) {
+        configuration.setClassLoader(getClassLoader(type, classLoader));
         try(final FSTObjectInput input = new FSTObjectInput(inputStream, configuration)) {
             @SuppressWarnings("unchecked") final T obj = (T)input.readObject(type);
             return obj;
@@ -56,7 +63,9 @@ public class FstSerde implements Serde {
         }
     }
 
-    @Override public <T> @Nullable T deserializeNullable(Class<T> type, InputStream inputStream) {
+    @Override
+    public <T> @Nullable T deserializeNullable(Class<T> type, InputStream inputStream, @Nullable ClassLoader classLoader) {
+        configuration.setClassLoader(getClassLoader(type, classLoader));
         try(final FSTObjectInput input = new FSTObjectInput(inputStream, configuration)) {
             @SuppressWarnings("unchecked") final @Nullable T obj = (T)input.readObject(type);
             return obj;
@@ -76,11 +85,25 @@ public class FstSerde implements Serde {
     }
 
     @Override
-    public @Nullable Object deserializeTypeAndObject(@Nullable ClassLoader classLoader, InputStream inputStream) {
+    public @Nullable Object deserializeObjectOfUnknownType(InputStream inputStream, @Nullable ClassLoader classLoader) {
+        configuration.setClassLoader(getClassLoader(classLoader));
         try(final FSTObjectInput input = new FSTObjectInput(inputStream, configuration)) {
             return input.readObject();
         } catch(Exception | IncompatibleClassChangeError e) {
             throw new DeserializeRuntimeException(e);
         }
+    }
+
+
+    private ClassLoader getClassLoader(Class<?> type, @Nullable ClassLoader classLoader) {
+        if(classLoader != null) return classLoader;
+        if(defaultClassLoader != null) return defaultClassLoader;
+        return getClassLoader(type.getClassLoader());
+    }
+
+    private ClassLoader getClassLoader(@Nullable ClassLoader classLoader) {
+        if(classLoader != null) return classLoader;
+        if(defaultClassLoader != null) return defaultClassLoader;
+        return getClass().getClassLoader();
     }
 }